chore: allow disabling per tokenizer (#136)
| Fixes RM-541 |
|:---:|

## 🧰 Changes

Allow fine grain control over which tokenizers are in use. While developing for @readme/editor, we'd like to keep the editor non-lossy. The easiest way to do that is to disable tokenizers that we don't yet handle.

If we don't disable a tokenizer, we could convert it back to plain text. But it gets complicated, because we'd have to know what text was 'eaten' by the markdown engine.