fix: magic tables collect all cells (#139)
| Fixes CX-188
|:---:

## 🧰 Changes

Previously, the `tableFlattening` function was assuming that tables only had plain text. In effect, if you have any markdown formatting in a table cell, it and anything after it would not be indexed!

## 🧬 QA & Testing

You can verify that the test is sane, but this really needs to be end to end
tested.