docs: remove unnecessary slug parameter, newlines
As detailed in this comment: https://github.com/readmeio/markdown/pull/173#issuecomment-867169250, the `slug` parameter in the frontmatter was only causing issues. `rdme` just looks at the filename to determine the slug, so the existence of the slug parameter here was just unnecessary duplicated information that has the potential to cause future syncing issues. I tested it locally with these changes and ran the command a bunch of times and confirmed that syncing works just fine without it.

Also I added newlines to the end of any files that were missing them!