refactor: adopt variable (#200)
[<img height=20 src=https://readme.com/static/favicon.ico align=center>][demo] |
:---:|

## 🧰 Changes

Adopts the `@readme/variable` package. From now on, you'll be able to import the `Variable`, `VARIABLE_REGEX`, and `SelectedAppContext` from `@readme/markdown`.

Note that `@readme/variable`'s old default export was the `Variable` component with the requisite context consumers. At the same time, it was exporting the `Variable` component as the named export: `Variable`. `@readme/markdown` will only be exporting the component wrapped with the consumers.

## TODO

There's a bit of a chicken and the egg type problem. This repo is depending on `@readme/syntax-highlighter` which is depending on `@readme/variable`. So we'll have to push an updated `@readme/markdown` first, before we can update `@readme/syntax-highlighter` to use `@readme/markdown`.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-200.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io