feat: save newlines with spaces (#471)
toolbox Changes
Override the break compiler to save as "␣␣\n" instead of "\\n".

The MarkdownEditor doesn't use the escape parser, so that the backslashes are preserved in the editor. For some reason remark-parse doesn't consume the backslash of "\\n" when parsing the break. (I believe this is technically incorrect, but shrug) This means that if you enter a newline in the editor, it gets saves as "\\n" and then loaded as the same.

We can't turn the escape parser on, as we'd lose escapes. And, we could save as "\n" when correctnewlines: false, but that's not portable to other markdown parsers. So this seems like a reasonable workaround.