chore: update deps (#496)
## 🧰 Changes

Updates a bunch of dependencies. We're stuck not updating any remark/unified until we can do a big rewrite. And we're also stuck not updating react until some other deps do too?

This update also as a `.npm-upgrade.json` file. I typically use `npm-upgrade` to do batch updates, so this is finally checking that file in.

This also changes the snapshot comparison from per pixel to SSIM. This was motivated by a new change in how the embeds thumbnail was being rendered. The thumbnail appears to be blurred vs pixelated depending on which node version it is. I believe this has more to do with puppeteer than any of the production code. Plus, I'd just like to see how this method works out.