feat: add safeMode flag (#517)
Adds a `safeMode` flag. When run with `safeMode: true`, we no longer render html blocks. This is to prevent a wide range of vulnerabilities that are available via iframes, script tags, and attributes that allow javascript.

There are cases where we want to render 'untrusted' markdown (via suggested edits). Since we want to still allow the full range of html within html blocks, but we also need to be able to render untrusted changes, this seemed like a good compromise.