feat: fix plain string compiler (#532)
Fixes the `astToPlainText` compiler.

I'm not sure how this working at all before, sometimes it returns a string, and sometimes it returns the passed in node?! Prior to this, it was calling `.runSync(ast)` which only [runs the `run` step](https://github.com/unifiedjs/unified#overview) and only calls `transformers` on the passed in `ast`. Theoretically, the correct call should be `.stringify(ast)`, which calls the `compile` step. 

This PR reworks `astToPlainText()` to make the 'correct' call to `.stringify(ast)`. It also copies in the `hast-util-to-string` library so that it will use a space (` `) to join all the strings together.