feat: don't render style in safeMode (#525)
Gives `style` tags the same treatment as html magic blocks. If you render with the option `safeMode`, we'll render `style` tags in a `pre` tag instead of as a `style` tag. The purpose is to prevent a phishing vector in suggested edits or discusssions.

It also strips `style` attributes when `safeMode` is on!

***

`safeMode` now prevents rendering:

- html magic blocks
- raw `style` tags
- `style` attributes anywhere

***

This PR touches so many files because of the way we were registering sanitation rules, it wasn't dynamic enough to be configurable. Incidentally, the `html` processor wasn't getting all the santitation rules from components. Which is now fixed as a result of these changes.

It also adds a 'Safe Mode' toggle to the demo app, which is now being used in some snapshot tests too!