chore: dockerize browser tests (#598)
Dockerize browser tests so people can run them locally.

The snapshot tests are pretty environment specific. CI typically uses ubuntu and most of the devs are running OSX. I believe a bunch of time was spent trying to get the font config to match, but it didn't quite work? So this re-opens an old PR that runs the snapshot tests in a docker container.

Prior to this, the workflow had been:

1. update a branch
2. wait for the CI to fail
3. download the snapshots/diffs from github
4. re-update branch

Now, if you're clever, you can run the tests before updating the branch. Or, more likely, not have to dig the github actions UI to download the snapshots, and instead generate them locally.

To update the snapshots:

```
make updateSnapshot
```