fix: treat code-tabs as div in hast (#664)
This allows putting code tab components within html tags.

For example:

````
<p>

\```json
const foo = 'bar'
\```
\```ruby
foo = :bar
\```

</p>
````

Previously, this did not work because `rehype-raw` won't treat a `code-tabs` node as proper html. It tries to recover, and splices the children of the `code-tabs` to be it's siblings. In the above example, you'd get a paragraph with 2 code blocks, rather than a single code tabs component. This is technically correct, so my workaround here, is to treat a `code-tabs` node as if it were a div. Then a custom `createElement` will render the proper `CodeTab` component.