test: browser tests (#778)
Browser tests!

Ok, so I'd kind of given up on maintaining the snapshot tests. But now that we're trying to add mdx, I _really_ want them. This PR adds a link to the latest app css bundles so there are lots of little changes to the styles. As well, we changed the layout of the demo app a bit, so I had to expand the viewport of the test browser.