feat: mdx-less reusable content (#811)
Adds support for Reusable Content!

This PR adds a transformer to convert `html` nodes into `reusable-content` nodes. If the so named content node has been passed in via `reusableContent`, then the parsed `mdsat` is inserted into place. This PR also adds a compiler for writing back out to `md`.

The syntax is meant to match html/JSX so we have compatibility with future MDX.