feat: single line magic blocks (#812)
Attempts to fix inserting magic blocks in container nodes.

Consider the following markdown:

```
> 👎🏼 doesn't work
>
> [block:image]
> {
>   "images": [
>     { ... }
>   ]
> }
> [/block]
```

The above does not parse as expected for a number of reasons. The main culprit is the default option to `normalize`. Normalizing magic blocks inserts newlines where there should **not** be ones. In this case separating the magic block from the callout.

Another thing that can happen around the edge cases, or if attempting to reorder the tokenizers, is that the JSON block's indentation could get parsed as indented code.

This PR does two things, removes the `normalize` option, and starts saving magic blocks onto a single line. The only downside I can see is that the resulting text is less readable. But, magic blocks are sort of an undocumented api anyway, and aren't typically edited directly anyway.