fix: table flattening (#830)
| [![PR App][icn]][demo] | Fix RM-8651 |
| :--------------------: | :---------: |

## 🧰 Changes

Fixes table flattening of html tables.

The prior code had the expectation that it would only be flattening markdown tables, so it only looked at the first 2 children of a table. This PR iterates over **all** the children of a table, and cleans up the output string a little bit.