style: fix glossary item (#890)
| [![PR App][icn]][demo] | Fix RM-9861 |
| :--------------------: | :---------: |

## 🧰 Changes

I changed the name of `GlossaryItem` to `Glossary`, but we never load
`@readme/mdx`'s css in the main app. This caused the glossary tooltip to
be unstyled.

I poked at getting rmdx's styles loaded, but ran into some issues with
codemirror. So for now, I'm relying on the old styling to still be
available, blheh.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg