feat: mdx TOC (#883)
| [![PR App][icn]][demo] | Fix RM-9179 |
| :--------------------: | :---------: |

## 🧰 Changes

TOC!!1!1

The old TOC worked by parsing the `mdast` for `heading` via
`mdast-util-toc`. This produces an `mdast` suitable for rendering via
the `TableOfContents` component. The package `remark-slug` appears to be
abandoned, and the [mdx
maintainers](https://github.com/orgs/mdx-js/discussions/2085) recommend
using `rehype-slug` and `rehype-extract-oc`. `rehype-extract-toc`
generates a POJO that represent the TOC, and it doesn't appear to have
any functions for parsing component imports. So, I've implemented a
plugin that parses the main doc and it's components `hast`'s during the
`compile` step. Then it can be rendered in the `TableOfContents` during
the `run` step. This changes the return signatures to be an object
instead of a single component, to prevent requiring 2 separate compile
and run steps when rendering.

Other notable changes:

- Refactored `run`, `compile`, `mdx`, `astProcessor` into their own
files under `./lib`
- Deleted the `Style` component.
- Deleted some tests (maybe not a great idea?).
- Added `execute` test helper.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg

---------

Co-authored-by: Rafe Goldberg <rafegoldberg@gmail.com>