fix: saving tables (#942)
| [![PR App][icn]][demo] | RM-9793 |
| :--------------------: | :-----: |

## 🧰 Changes

Fixes saving 'complex' tables.

In a prior PR, I added a new node type `tableau` to represent tables
with block content. This mostly worked for the editor, but not when
doing `mdx(mdast(doc))`. This PR adds a transformer, so the a `tableau`
is converted to a JSX table just like regular `table`'s.

This also adds the check so, if a `tableau` no longer has block content,
it's saved as markdown!

## 🧬 QA & Testing

Hard to test outside the main app.

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg