fix: legacy image migration (#975)
[![PR App][icn]][demo] | Ref CX-1113, CX-1114
:-------------------:|:----------:

## 🧰 Changes

Fixes migrating images with legacy data.

We deprecated a particular format of image magic block, but we failed to
ever migrate pages away from it. They've been broken/changed for awhile,
and I'm not sure if we could correctly detect the format at this point?

When they get parsed into an AST, they appear to be inline image nodes,
which can then compile incorrectly. We've been treating image nodes that
are alone in a paragraph as block images for awhile, so lets encode that
in the migration.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg