fix: image border migration (#13027)
| 📖 [PR App][pr] | 🎫 Resolve RM-11043 |
| :-------------: | :-----------------: |

## 🧰 Changes

Fixes migrating images with borders.

I'm not clear on the timeline/cause, but magic blocks images used `{
className: 'border' }` for setting the border. The mdx images use:
```
<Image border={true} />
```

I think this is clearly a nicer API, but I'm not sure how we missed
adding this to the migration? Or maybe it's somehow regressed??? The
editor portion is definitely a regression from me refactoring images to
fix captions.

## 💁 Customer Impact

<!-- PLEASE DESCRIBE IF THIS CHANGE AFFECTS CUSTOMERS AND HOW -->
<!-- This is extremely useful for Tony and Marc writing the changelog!
-->

## 🧬 QA & Testing

- [Broken on next][next].
- [Working in this PR][pr]!

[next]: https://next.readme.ninja
[pr]: https://readme-pr-13027.readme.ninja
[ui]: https://readme-pr-13027.readme.ninja/ui

<!-- Uncomment and unescape this if you don't want a PR app! -->
<!-- \[skip preview\] -->