fix: normalizing the doc (#13047)
| 📖 [PR App][pr] | 🎫 Resolve RM-11134 |
| :-------------: | :-----------------: |

## 🧰 Changes

Fixes a migration regression.

When I started added migration specific transforms to the main app, I
inadvertantly bypassed a normalization step that RDMD usually applies.
There are magic blocks in the wild that don't have enough newlines, and
get incorrectly parsed by RDMD. So, RDMD, usually does a simple search
and replace. This was bypassed in the migration, when I switch to using
`rdmd.processor()`. Since it doesn't operate on the doc, it doesn't
apply the normalization step.

## 💁 Customer Impact

<!-- PLEASE DESCRIBE IF THIS CHANGE AFFECTS CUSTOMERS AND HOW -->
<!-- This is extremely useful for Tony and Marc writing the changelog!
-->

## 🧬 QA & Testing

- [Broken on next][next].
- [Working in this PR][pr]!

[next]: https://next.readme.ninja
[pr]: https://readme-pr-13047.readme.ninja
[ui]: https://readme-pr-13047.readme.ninja/ui

<!-- Uncomment and unescape this if you don't want a PR app! -->
<!-- \[skip preview\] -->