fix: make mermaid optional (#999)
[![PR App][icn]][demo] |
:-------------------:|

## 🧰 Changes

Moves `mermaid` to an optional dependency.

Adding `mermaid` appears to have increased the time to import the dist
by a factor of almost 2. This is mostly a concern in our backend
services, ie, `mdx-renderer` and `gitto`. This is also completely
unnecessary since `mermaid` does not easily support SSR. (We'd have to
get a working puppeteer setup to render it on the server, which is
probably not worth the effort at the moment.)

Also, I fixed up how `mermaid` class names were being applied in code
tabs.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg