feat: compat parser (#1004)
[![PR App][icn]][demo] |
:-------------------:|

## 🧰 Changes

Moves the migration bits back to here.

This PR attempts to pull in the changes from the main readme app that
are specific to the on going migration to MDX. We had originally started
moving migration specific changes to the main app for speed of
development, and organization. But now, we need to be able to move the
migration bits into the gitto backend. I tried to preserve the git
histories of the specific files, so there's a little more context.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg