fix: migrate tables with html (#1005)
[![PR App][icn]][demo] | CX-1192
:-------------------:|:----------:

## 🧰 Changes

Fixes migrating tables with html.

We're doing some shenanigans with table cells so that they support block
content. We basically reparse each cell as if it's a root markdown doc
and then re-insert it into the tree.

This adds block level html as a condition for wrapping said content in a
paragraph. Wrapping it in a paragraph seems to be required to get remark
to compile newlines correctly. :shrug:


## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg