fix: dont treat escapes as flow (#1040)
[![PR App][icn]][demo] | Ref CX-1610
:-------------------:|:----------:

## 🧰 Changes

Fixes migrating tables with leading escapes.

The `phrasing` util is not aware of the `escape` node type, as it is
from a previous version of `mdast`. This was confusing our
`tables-to-jsx` transformer. It would look at tables and decide if it
could be compiled to markdown format, eg:
```
| a | b |
|---|---|
| 1 | 2 |
```
or if it should be compiled to JSX format, eg:
```jsx
<Table>
  <TableHead>
    <TableRow>
      <TableCell>a</TableCell>
      <TableCell>b</TableCell>
    </TableRow>
  </TableHead>
  <TableBody>
    <TableRow>
      <TableCell>1</TableCell>
      <TableCell>2</TableCell>
    </TableRow>
  </TableBody>
</Table>
```

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg