fix: nesting children (#1060)
| [![PR App][icn]][demo] | Ref RM-12072 |
| :--------------------: | :----------: |

## 🧰 Changes

Fixes a crash when rendering exported components.

The error was caused because we were passing in an map of custom block
tags (not exports) to `tocToMdx`. The `tocToMdx` function would parse
the toc for tags, and see if they had an associated `toc` export. By for
exported components, that (`component[node.name]`) would be `undefined`,
and we would crash. This refactors it to be slightly clearer and less
error prone.

## ✅ TODO

**How do we handle generating a Table of Contents for Custom
Components??**

We currently support parsing a table of contents from the default
export, because that's how reusable content works. This is probably not
the best way to do it with regards to custom components. The default, or
any of the exports, could be JSX or markdown or any kinda of ESM object.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg