fix: loading styles in a shadow-dom (#1064)
| [![PR App][icn]][demo] | Ref RM-12182, RM-12042 |
| :--------------------: | :--------------------: |

## 🧰 Changes

Updates `TailwindStyle` to work in a shadow root.

This PR reworks the `TailwindStyle` component to not push styles into
the `head`. It also adds back generating the stylesheet on load, as the
timing appears to be different in the main app than the demo.

## 🧬 QA & Testing

I haven't been able to get `npm link` to work recently. Copying the dist
might, but I was getting build errors, so I wasn't confident that it was
working properly.

That being said, I copied the `TailwindStyle` component directly into
the app to dev and test.

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg