fix: use data attribute for dark theme (#1066)
| [![PR App][icn]][demo] | Ref RM-12042 |
| :--------------------: | :----------: |

## 🧰 Changes

Updates the tailwind stylesheet to support dark mode via a data
attribute.

This will allow us the main app to toggle between light and dark mode in
the component editor preview. To use it, you pass in
`darkModeDataAttribute` to the `TailwindStyle` component, and then add a
data attribute (`data-theme="dark"`) somewhere in the document above the
mdx content.

This isn't exactly represented in the demo app. I did add a toggle so
you can switch from dark mode system support to data attribute support.
And then this PR adds a component that uses it's own `data-theme`
attribute.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg