fix: allow configurable theme (#1069)
| [![PR App][icn]][demo] | Ref RM-12311 |
| :--------------------: | :----------: |

## 🧰 Changes

Allows the tailwind dark mode attribute to be configurable.

My last attempts were too hasty. Prior to this, if there was no data
attribute param, dark mode would fall back to the system preference.
Then you could turn on using the data attribute, which would be the same
as the readme app's. This works for the standard view mode. But, in the
preview, we want it to be seperately configurable from main theme.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg