fix: pass through theme prop to CodeTabs from RunOpts (#1072)
[![PR App][icn]][demo] | Fix RM-12348
:-------------------:|:----------:

## 🧰 Changes
We weren't including the `theme` in `RunOpts`, and no props are being
exclusively passed in to the components. This meant Mermaid Diagrams
would always fallback to the `default` theme and not appropriately match
the project theme.

I added a `ThemeContext` through which we can set and access the project
theme that's passed in via the `RunOpts`.

## 🧬 QA & Testing

Note: I can't figure out how to test these changes locally linked with
the `readme` repo 😭 . Without sinking too much time in it, may just want
to try to get this published in order to test via `readme`.

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg