fix: allow missing components (#1076)
| [![PR App][icn]][demo] | Ref RM-12219 |
| :--------------------: | :----------: |

## 🧰 Changes

Adds an option to either or ignore or throw if there are missing
components.

- `missingComponents === "ignore"`
This is useful for when rendering a page that has missing components. We
don't want to get into this state, but it's definitely possible, so we
should fallback gracefully.
- `missingComponents === "throw"`
We still need to be able to quickly tell if a page is in a bad state.
And we want it to be a much faster check than doing a full render.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg