feat: external large deps (#1091)
| [![PR App][icn]][demo] |
| :--------------------: |

## 🧰 Changes

Externalizes `@readme/syntax-highlighter` and `acorn`.

Those deps are duplicated across the main readme app, and definitely do
not need to be bundled.

It also removes an import of `react-dom/server` which seems fairly
extraneous. The `HTMLBlock` component was checking if you passed in
react components as children, and converting them to a string. I don't
think that's something we should support.

### Webpack Bundle Analyzer

**Before** (parsed size of 4.05MB)

![image](https://github.com/user-attachments/assets/83dccdb6-920b-4239-aabf-6392356655fa)

**After** (parsed size of 2.5MB)

![image](https://github.com/user-attachments/assets/e5dddf90-c7dc-49ae-bbb1-65c399333670)



## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg