refactor: callout headings (#1123)
| [![PR App][icn]][demo] | Ref RM-12848 |
| :--------------------: | :--------: |

## 🧰 Changes

Removes adding `h3`'s to callouts by default, when using JSX directly.

This will leave the behavior of using rdmd style callouts the same. The
following will still add an `h3`:
```
> 📘 Info
>
> This is the info callout.
```

But, if you use JSX directly, it will not add an `h3` by default. You
could of course add the `h3`:
```
<Callout theme="info">
### Info

This is the info callout.
</Callout>
```

---
This will allow the editor to promote not adding an `h3` by default and
still be backwards compatible.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg

---------

Co-authored-by: Tony Li <runnabro@users.noreply.github.com>