fix: update glossary to use tooltip with mutation observer (#1173)
[![PR App][icn]][demo] | Fix CX-2117
:-------------------:|:----------:

## 🧰 Changes

- Creates a `Tooltip` component that is a wrapper around
`@tippyjs/react` and attaches a mutation observer to the trigger element
- Update `Glossary` to use the new `Tooltip` so that when something like
translations blows away its DOM tree, it can get an updated reference to
the new trigger node


## 🧬 QA & Testing

Tried replicating this in the local `markdown` app since we can't `npm
link` this package to the main app. I will check the fix again in my
local after a new markdown version is published!

### Before


https://github.com/user-attachments/assets/45015a35-db0c-4b36-8af3-a6560be145ea

### After


https://github.com/user-attachments/assets/984ed7df-c43b-413d-b5f5-d57800dcebe0

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg