fix: react hydration mismatches (#1198)
[![PR App][icn]][demo] | Fix CX-2284
:-------------------:|:----------:

We've got a few markdown components that are causing hydration
mismatches when they get rendered in the Hub.
This PR patches them up!

## 🧰 Changes
1. Wait until after hydration to apply syntax highlighting to
the`<Code>` component
2. Prevent mermaid's default initialization on module load from racing
with the client side render in `<CodeTabs>`

I tried to comb through all of our custom blocks to find any others that
cause these errors and only found the code blocks, mermaid charts and
custom components causing issues. Fix for the custom components
hydration is [up in this
PR](https://github.com/readmeio/readme/pull/15890) on the monorepo.


## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].


[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg

---------

Co-authored-by: Kelly Joseph Price <kellyjosephprice@gmail.com>