fix: pipes in tables (#1205)
| [![PR App][icn]][demo] | Fix CX-2392 |
| :--------------------: | :---------: |

## 🧰 Changes

Fixes a regression in our markdown editor that would not properly escape
pipes (`|`) in tables.

This allows users to write pipe characters directly in tables without
them being interpreted as column separators. Normally, I would put this
kind of convenience directly in the editor code. However, the logic to
compile a table to gfm or jsx is in this library, and we want the
escpaing to happen after the jsx conversion. That way we're not
erronously escaping jsx tables.

## 🧬 QA & Testing

- [Broken on production][prod].
- [Working in this PR app][demo].

[demo]: https://markdown-pr-PR_NUMBER.herokuapp.com
[prod]: https://SUBDOMAIN.readme.io
[icn]:
https://user-images.githubusercontent.com/886627/160426047-1bee9488-305a-4145-bb2b-09d8b757d38a.svg